Fix cut and paste error, read the long value not the integer value
When reading the underlying UnsignedLong value grab the long, not the
int

+++ proton-j/src/main/java/org/apache/qpid/proton/codec/UnsignedLongType.java
@@ -61,7 +61,7 @@ public class UnsignedLongType extends AbstractPrimitiveType<UnsignedLong>
 
     public void fastWrite(EncoderImpl encoder, UnsignedLong value)
     {
-        long longValue = value.intValue();
+        long longValue = value.longValue();
         if (longValue == 0)
         {
             encoder.writeRaw(EncodingCodes.ULONG0);
